+++ packages/ub/_autotest/020_selectForAllEntity.js
@@ -5,26 +5,20 @@
  */
 const assert = require('assert')
 const ok = assert.ok
-const fs = require('fs')
 const cmdLineOpt = require('@unitybase/base').options
 const argv = require('@unitybase/base').argv
 const TEST_NAME = 'Select from all entities'
 
 module.exports = function runTest (options) {
-  let session, conn
-
   if (!options) {
-    let opts = cmdLineOpt.describe('', TEST_NAME)
+    const opts = cmdLineOpt.describe('', TEST_NAME)
       .add(argv.establishConnectionFromCmdLineAttributes._cmdLineParams)
     options = opts.parseVerbose({}, true)
     if (!options) return
   }
 
-  session = argv.establishConnectionFromCmdLineAttributes(options)
-  // if (!session.__serverStartedByMe) {
-  //   throw new Error('Shut down server before run this test')
-  // }
-  conn = session.connection
+  const session = argv.establishConnectionFromCmdLineAttributes(options)
+  const conn = session.connection
 
   try {
     console.debug('Start ' + TEST_NAME)
@@ -40,11 +34,17 @@ module.exports = function runTest (options) {
   function testAllSelect (conn) {
     let recCnt = 0
     let ettCnt = 0
-    let domain = conn.getDomainInfo()
+    const domain = conn.getDomainInfo()
     domain.eachEntity(function (entity, eName) {
       if (entity.haveAccessToMethod('select')) {
         console.debug('run select for %s', eName)
-        let res = conn.Repository(eName).attrs(Object.keys(entity.attributes)).limit(10).selectAsArray()
+        const defaultViewAttrs = []
+        entity.eachAttribute((attr, name) => {
+          if (attr.defaultView) {
+            defaultViewAttrs.push(name)
+          }
+        })
+        const res = conn.Repository(eName).attrs(defaultViewAttrs).limit(10).selectAsArray()
         ok(typeof res === 'object' && res.resultData &&
           res.resultData.data && Array.isArray(res.resultData.data) &&
           res.resultData.rowCount === +res.resultData.rowCount, 'result is dataStore in array representation. Entity: ' + eName)
@@ -57,4 +57,3 @@ module.exports = function runTest (options) {
     console.debug('Entity tested %d. Rows selected %d', ettCnt, recCnt)
   }
 }
-